Refactor log stuff. Added more action log.
