+++ wildlifelicensing/apps/applications/templates/wl/conditions/assessor_enter_conditions.html
@@ -16,23 +16,17 @@
 {% block requirements %}
     require([
         'js/conditions/assessor_enter_conditions',
-        'js/communications_log',
-        'js/actions_log
-    ], function (enter_conditions, commsLog, actionsLog) {
+        'js/logs'
+    ], function (enter_conditions, logs) {
         $(function () {
             enter_conditions.init({{ assessment|jsonify }}, {{ application|jsonify }},
                                   {{ form_structure|jsonify }}, {{ other_assessments|jsonify }});
 
-            commsLog.initCommunicationLog({
-                showLogPopoverSelector: '#showLog',
-                showLogEntryModalSelector: '#addLogEntry',
-                logEntryModalSelector: '#logEntryModal',
-                logEntryFormSelector: '#addLogEntryForm',
+            logs.initCommunicationLog({
                 logListURL: "{% url 'wl_applications:log_list' application.id %}",
                 addLogEntryURL: "{% url 'wl_applications:add_log_entry' application.id %}"
             });
-            actionsLog.init({
-                showLogPopoverSelector: '#showActionLog',
+            logs.initActionLog({
                 logListURL: "{% url 'wl_applications:action_list' application.id %}",
             });
 
@@ -112,9 +106,7 @@
                         </div>
                     </div>
                 </div>
-                {% with disable_collapse=True add_text='Add application log entry' %}
-                    {% include 'wl/communications_panel.html' %}
-                {% endwith %}
+                {% include 'wl/logs_panel.html' %}
                 <div class="row">
                     <div class="col-md-12">
                         <div class="panel panel-default">
@@ -185,5 +177,5 @@
 {% endblock %}
 
 {% block modals %}
-    {% include 'wl/communications_modal.html' %}
+    {% include 'wl/logs_comm_entry_modal.html' %}
 {% endblock %}
