+++ wildlifelicensing/apps/applications/templates/wl/conditions/enter_conditions.html
@@ -14,27 +14,22 @@
     <link href="{% static 'wl/css/application.css' %}" rel="stylesheet"/>
 {% endblock %}
 
+
 {% block requirements %}
     require([
         'js/conditions/enter_conditions',
-        'js/communications_log',
-        'js/actions_log'
-    ], function (enterConditions, commLog, actionsLog) {
-           enterConditions.init({{ application|jsonify }}, {{ assessments|jsonify }}, {{form_structure|jsonify}}, '{{ csrf_token }}');
+        'js/logs'
+    ],
+    function (enterConditions, logs) {
+        enterConditions.init({{ application|jsonify }}, {{ assessments|jsonify }}, {{form_structure|jsonify}}, '{{ csrf_token }}');
 
-            commLog.initCommunicationLog({
-                showLogPopoverSelector: '#showLog',
-                showLogEntryModalSelector: '#addLogEntry',
-                logEntryModalSelector: '#logEntryModal',
-                logEntryFormSelector: '#addLogEntryForm',
-                logListURL: "{% url 'wl_applications:log_list' application.id %}",
-                addLogEntryURL: "{% url 'wl_applications:add_log_entry' application.id %}"
-            });
-            actionsLog.init({
-                showLogPopoverSelector: '#showActionLog',
-                logListURL: "{% url 'wl_applications:action_list' application.id %}",
+        logs.initCommunicationLog({
+            logListURL: "{% url 'wl_applications:log_list' application.id %}",
+            addLogEntryURL: "{% url 'wl_applications:add_log_entry' application.id %}"
         });
-
+        logs.initActionLog({
+            logListURL: "{% url 'wl_applications:action_list' application.id %}",
+        });
     });
 {% endblock %}
 
@@ -117,9 +112,7 @@
                         </div>
                     </div>
                 </div>
-                {% with disable_collapse=True add_text='Add application log entry' %}
-                    {% include 'wl/communications_panel.html' %}
-                {% endwith %}
+                {% include 'wl/logs_panel.html' %}
                 <div class="row">
                     <div class="col-md-12">
                         <div class="panel panel-default">
@@ -250,5 +243,5 @@
             </div>
         </div>
     </div>
-    {% include 'wl/communications_modal.html' %}
+    {% include 'wl/logs_comm_entry_modal.html' %}
 {% endblock %}
