+++ wildlifelicensing/apps/applications/templates/wl/issue/issue_licence.html
@@ -16,18 +16,18 @@
 {% endblock %}
 
 {% block requirements %}
-    require(['js/issue/issue_licence', 'js/communications_log'], function(issueLicence, commsLog) {
-        $(function () {
-            issueLicence.initialise();
-
-            commsLog.initCommunicationLog({
-                showLogPopoverSelector: '#showLog',
-                showLogEntryModalSelector: '#addLogEntry',
-                logEntryModalSelector: '#logEntryModal',
-                logEntryFormSelector: '#addLogEntryForm',
-                logListURL: "{% url 'wl_applications:log_list' application.id %}",
-                addLogEntryURL: "{% url 'wl_applications:add_log_entry' application.id %}"
-            });
+    require([
+        'js/issue/issue_licence',
+        'js/logs'
+    ],
+    function(issueLicence, logs) {
+        issueLicence.initialise();
+        logs.initCommunicationLog({
+            logListURL: "{% url 'wl_applications:log_list' application.id %}",
+            addLogEntryURL: "{% url 'wl_applications:add_log_entry' application.id %}"
+        });
+        logs.initActionLog({
+            logListURL: "{% url 'wl_applications:action_list' application.id %}",
         });
     });
 {% endblock %}
@@ -116,9 +116,7 @@
                         </div>
                     </div>
                 </div>
-                {% with add_text='Add application log entry' %}
-                    {% include 'wl/communications_panel.html' %}
-                {% endwith %}
+                {% include 'wl/logs_panel.html' %}
                 {% if payment_status == 'Awaiting Payment' %}
                     <button id="issue" type="button" class="btn btn-sm btn-success pull-right disabled"
                             data-title="Payment must be received before licence can be issued">
@@ -154,5 +152,5 @@
 {% endblock %}
 
 {% block modals %}
-    {% include 'wl/communications_modal.html' %}
+    {% include 'wl/logs_comm_entry_modal.html' %}
 {% endblock %}
