+++ wildlifelicensing/apps/applications/templates/wl/process/process_app.html
@@ -15,27 +15,21 @@
 {% endblock %}
 
 {% block requirements %}
-    require(['js/process/process_app', 'js/communications_log', 'js/actions_log'],
-        function (app, commsLog, actionsLog) {
+    require(['js/process/process_app', 'js/logs'],
+        function (app, logs) {
             $(function () {
                 data = {{ data|jsonify }};
                 app.initialiseApplicationProcesssing(data);
 
-                commsLog.initCommunicationLog({
-                showLogPopoverSelector: '#showLog',
-                showLogEntryModalSelector: '#addLogEntry',
-                logEntryModalSelector: '#logEntryModal',
-                logEntryFormSelector: '#addLogEntryForm',
-                logListURL: "{% url 'wl_applications:log_list' data.application.id %}",
-                addLogEntryURL: "{% url 'wl_applications:add_log_entry' data.application.id %}"
-            });
+                logs.initCommunicationLog({
+                    logListURL: "{% url 'wl_applications:log_list' data.application.id %}",
+                    addLogEntryURL: "{% url 'wl_applications:add_log_entry' data.application.id %}"
+                });
+                logs.initActionLog({
+                    logListURL: "{% url 'wl_applications:action_list' data.application.id %}",
+                });
 
-            actionsLog.init({
-                showLogPopoverSelector: '#showActionLog',
-                logListURL: "{% url 'wl_applications:action_list' data.application.id %}",
-            });
-
-            $('#mainContainer').removeClass('hidden');
+                $('#mainContainer').removeClass('hidden');
             });
         });
 {% endblock %}
@@ -125,7 +119,7 @@
                             </div>
                         </div>
                     </div>
-                    {% include 'wl/communications_panel.html' %}
+                    {% include 'wl/logs_panel.html' %}
                     <div class="row">
                         <div class="col-md-12">
                             <div class="panel panel-default">
@@ -462,6 +456,6 @@
                 </div>
             </div>
         </div>
-        {% include 'wl/communications_modal.html' %}
+        {% include 'wl/logs_comm_entry_modal.html' %}
     {% endblock %}
 {% endblock %}
