+++ wildlifelicensing/apps/applications/templates/wl/view/view_readonly.html
@@ -14,18 +14,21 @@
 {% endblock %}
 
 {% block requirements %}
-    require(['jQuery', 'js/entry/application_preview', 'js/communications_log'], function($, applicationPreview, commsLog) {
+    require([
+        'jQuery',
+        'js/entry/application_preview',
+        'js/logs'
+    ],
+    function($, applicationPreview, logs) {
         applicationPreview.layoutPreviewItems('#applicationContainer', {{ application.licence_type.application_schema|jsonify }}, {{ application.data|jsonify }});
-
         $('#mainContainer').removeClass('hidden');
 
-        commsLog.initCommunicationLog({
-            showLogPopoverSelector: '#showLog',
-            showLogEntryModalSelector: '#addLogEntry',
-            logEntryModalSelector: '#logEntryModal',
-            logEntryFormSelector: '#addLogEntryForm',
+        logs.initCommunicationLog({
             logListURL: "{% url 'wl_applications:log_list' application.id %}",
             addLogEntryURL: "{% url 'wl_applications:add_log_entry' application.id %}"
+        });
+        logs.initActionLog({
+            logListURL: "{% url 'wl_applications:action_list' application.id %}",
         });
 
         // need to initialise sidebar menu after showing main container otherwise affix height will be wrong
@@ -64,9 +67,7 @@
             <div class="col-md-3">
                 <nav class="hidden-xs top-buffer">
                     {% if customer %}
-                        {% with disable_collapse=True add_text='Add entry' %}
-                            {% include 'wl/communications_panel.html' %}
-                        {% endwith %}
+                        {% include 'wl/logs_panel.html' %}
                     {% else %}
                         <div class="row">
                             <div class="col-md-12">
@@ -103,6 +104,6 @@
 {% endblock %}
 {% block modals %}
  	{% if customer %}
-    	{% include 'wl/communications_modal.html' %}
+    	{% include 'wl/logs_comm_entry_modal.html' %}
    	{% endif %}
 {% endblock %}
