+++ wildlifelicensing/apps/applications/templates/wl/view/view_readonly_assessor.html
@@ -14,18 +14,16 @@
 {% endblock %}
 
 {% block requirements %}
-    require(['js/view/view_readonly_assessor', 'js/communications_log'], function (conditions, commsLog) {
+    require(['js/view/view_readonly_assessor', 'js/logs'], function (conditions, logs) {
         $(function () {
             conditions.init({{ assessment|jsonify }}, {{ application|jsonify }},
                             {{ form_structure|jsonify }}, {{ other_assessments|jsonify }});
-
-            commsLog.initCommunicationLog({
-                showLogPopoverSelector: '#showLog',
-                showLogEntryModalSelector: '#addLogEntry',
-                logEntryModalSelector: '#logEntryModal',
-                logEntryFormSelector: '#addLogEntryForm',
+            logs.initCommunicationLog({
                 logListURL: "{% url 'wl_applications:log_list' application.id %}",
                 addLogEntryURL: "{% url 'wl_applications:add_log_entry' application.id %}"
+            });
+            logs.initActionLog({
+                logListURL: "{% url 'wl_applications:action_list' application.id %}",
             });
         });
     });
@@ -103,9 +101,7 @@
                         </div>
                     </div>
                 </div>
-                {% with disable_collapse=True add_text='Add application log entry' %}
-                    {% include 'wl/communications_panel.html' %}
-                {% endwith %}
+                {% include 'wl/logs_panel.html' %}
                 <a href="{% url 'home' %}" type="button" class="btn btn-sm btn-primary pull-right">Close</a>
             </div>
             <div class="col-md-9">
@@ -149,5 +145,5 @@
 {% endblock %}
 
 {% block modals %}
-    {% include 'wl/communications_modal.html' %}
+    {% include 'wl/logs_comm_entry_modal.html' %}
 {% endblock %}
