+++ wildlifelicensing/apps/customer_management/templates/wl/customer_lookup.html
@@ -311,6 +311,6 @@
 
 {% block modals %}
     {% if customer %}
-        {% include 'wl/communications_modal.html' %}
+        {% include 'wl/logs_comm_entry_modal.html' %}
     {% endif %}
 {% endblock %}
