+++ wildlifelicensing/apps/main/static/wl/js/logs.js
@@ -7,10 +7,10 @@ define(['jQuery', 'lodash', 'moment', 'js/wl.dataTable'], function ($, _, moment
     function initCommunicationLog(options) {
         // default options
         options = _.defaults(options || {}, {
-            showLogPopoverSelector: null,
-            showLogEntryModalSelector: null,
-            logEntryModalSelector: null,
-            logEntryFormSelector: null,
+            showLogPopoverSelector: '#showCommunicationLog',
+            showLogEntryModalSelector: '#addCommunicationLogEntry',
+            logEntryModalSelector: '#logCommEntryModal',
+            logEntryFormSelector: '#addLogCommEntryForm',
             logTableSelector: $('<table id="communicationsLog-table" class="table table-striped table-bordered dataTable">'),
             logListURL: 'insert-default-url-here',
             addLogEntryURL: 'insert-default-url-here'
@@ -35,7 +35,7 @@ define(['jQuery', 'lodash', 'moment', 'js/wl.dataTable'], function ($, _, moment
         }
 
         // init log table
-        logDataTable = initLogTable(options.logListURL, options.logTableSelector);
+        logDataTable = initCommunicationTable(options.logListURL, options.logTableSelector);
 
         // init log table popover if provided
         if (options.showLogPopoverSelector) {
@@ -87,7 +87,7 @@ define(['jQuery', 'lodash', 'moment', 'js/wl.dataTable'], function ($, _, moment
                             $logEntryModal.modal('hide');
 
                             // clear/reset form fields
-                            $(this).find('#id_type').val($('#id_type option:first').val());
+                            $(this).find('#id_type').val($('#id_type').find('option:first').val());
                             $(this).find('#id_subject').val('');
                             $(this).find('#id_text').val('');
                             $(this).find('#id_attachment').val('');
@@ -99,7 +99,7 @@ define(['jQuery', 'lodash', 'moment', 'js/wl.dataTable'], function ($, _, moment
         }
     }
 
-    function initLogTable(logListURL, tableSelector) {
+    function initCommunicationTable(logListURL, tableSelector) {
         function commaToNewline(s){
              return s.replace(/[,;]/g, '\n');
         }
@@ -236,7 +236,108 @@ define(['jQuery', 'lodash', 'moment', 'js/wl.dataTable'], function ($, _, moment
         return dataTable.initTable($table, tableOptions, colDefinitions);
     }
 
+    function initActionLog(options) {
+        // multi-used selectors
+        var $logListContent, logDataTable;
+
+        // default options
+        options = _.defaults(options || {}, {
+            showLogPopoverSelector: '#showActionLog',
+            logTableSelector: $('<table id="actionsLog-table" class="table table-striped table-bordered dataTable">'),
+            logListURL: 'insert-default-url-here'
+        });
+
+        // if log table is in a popover, need to prepare log table container before initializing table or
+        // search/paging/etc won't show
+        if (options.showLogPopoverSelector) {
+            $logListContent = $('<div>').append($(options.logTableSelector));
+        }
+
+        // init log table
+        logDataTable = initActionTable(options.logListURL, options.logTableSelector);
+
+        // init log table popover if provided
+        if (options.showLogPopoverSelector) {
+            $(options.showLogPopoverSelector).popover({
+                container: 'body',
+                title: 'Action log',
+                content: $logListContent,
+                placement: 'right',
+                trigger: "manual",
+                html: true
+            }).click(function () {
+                var isVisible = $(this).data()['bs.popover'].tip().hasClass('in');
+                if (!isVisible) {
+                    logDataTable.ajax.reload();
+                    $(this).popover('show');
+                    $('[data-toggle="tooltip"]').tooltip();
+                } else {
+                    $(this).popover('hide');
+                }
+            });
+        }
+    }
+
+    function initActionTable(logListURL, tableSelector) {
+        var $table = $(tableSelector),
+            tableOptions = {
+                paging: true,
+                info: true,
+                searching: true,
+                processing: true,
+                deferRender: true,
+                serverSide: false,
+                autowidth: true,
+                order: [[2, 'desc']],
+                // TODO: next one is to avoid the 'search' field to go out of the popover (table width is small).
+                // see https://datatables.net/reference/option/dom
+                dom:
+                "<'row'<'col-sm-5'l><'col-sm-6'f>>" +
+                "<'row'<'col-sm-12'tr>>" +
+                "<'row'<'col-sm-5'i><'col-sm-7'p>>",
+                ajax: {
+                    url: logListURL
+                }
+            },
+            colDefinitions = [
+                {
+                    title: 'Who',
+                    data: 'who'
+                },
+                {
+                    title: 'What',
+                    data: 'what'
+                },
+                {
+                    title: 'When',
+                    data: 'when',
+                    render: function (date) {
+                        return moment(date).format(DATE_TIME_FORMAT);
+                    }
+                }
+            ];
+
+        // set DT date format sorting
+        dataTable.setDateTimeFormat(DATE_TIME_FORMAT);
+
+        // activate popover when table is drawn.
+        $table.on('draw.dt', function () {
+            var $tablePopover = $table.find('[data-toggle="popover"]');
+            if ($tablePopover.length > 0) {
+                $tablePopover.popover();
+                // the next line prevents from scrolling up to the top after clicking on the popover.
+                $($tablePopover).on('click', function (e) {
+                    e.preventDefault();
+                    return true;
+                });
+            }
+        });
+
+        return dataTable.initTable($table, tableOptions, colDefinitions);
+    }
+
     return {
-        initCommunicationLog: initCommunicationLog
+        initCommunicationLog: initCommunicationLog,
+        initActionLog: initActionLog
     };
 });
